refactor: dissolve caller-side replay ownership
