+++ src/cli/commands/router-types.ts
@@ -1,7 +1,7 @@
 import type { CliFlags } from '@agent-device/contracts/command';
 import type { AgentDeviceClient } from '../../agent-device-client.ts';
 import type { CliCommandName } from '../../command-catalog.ts';
-import type { ReplayTestReporterRuntime } from '../../replay/test/reporting.ts';
+import type { ReplayTestReporterRuntime } from '../replay-test/reporting.ts';
 
 export type ClientCommandParams = {
   positionals: string[];
