+++ src/cli/replay-test/__tests__/session-test-reporter-values.test.ts
@@ -13,7 +13,7 @@
 import { expect, test, vi } from 'vitest';
 import { mkdtempForTestSync } from '../../../__tests__/test-utils/tmp-dir.ts';
 
-vi.mock('../snapshot-interactor-capture.ts', () => ({
+vi.mock('../../../daemon/handlers/snapshot-interactor-capture.ts', () => ({
   captureSnapshotWithInteractor: vi.fn(async () => {
     throw new Error('no device runner available in this test');
   }),
@@ -23,9 +23,9 @@ import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 import type { ReplaySuiteResult } from '@agent-device/contracts/replay';
-import { handleSessionCommands } from './session-command-harness.ts';
-import { SessionStore } from '../../session-store.ts';
-import type { DaemonRequest, DaemonResponse } from '../../types.ts';
+import { handleSessionCommands } from '../../../daemon/handlers/__tests__/session-command-harness.ts';
+import { SessionStore } from '../../../daemon/session-store.ts';
+import type { DaemonRequest, DaemonResponse } from '../../../daemon/types.ts';
 import { withRequestProgressSink } from '@agent-device/host-kit/request';
 import { withTestDeviceInventoryProvider as withDeviceInventoryProvider } from '../../../__tests__/test-utils/device-inventory-gateways.ts';
 import type { DeviceInfo } from '@agent-device/kernel/device';
@@ -33,11 +33,8 @@ import {
   getReplayTestReporterExitCode,
   runReplayTestReporterProgress,
   runReplayTestReporters,
-} from '../../../replay/test/reporters/registry.ts';
-import type {
-  ReplayTestReporter,
-  ReplayTestReporterContext,
-} from '../../../replay/test/reporters/types.ts';
+} from '../reporters/registry.ts';
+import type { ReplayTestReporter, ReplayTestReporterContext } from '../reporters/types.ts';
 
 type RecordedHook = {
   hook: 'onSuiteStart' | 'onTestStart' | 'onTestStep' | 'onTestResult' | 'onSuiteEnd';
