+++ src/cli/replay-test/reporters/__tests__/progress.test.ts
