+++ src/cli/replay-test/reporters/__tests__/registry.test.ts
