+++ src/cli/replay-test/reporters/registry.ts
