+++ src/daemon/handlers/__tests__/session-replay-script-source.test.ts
@@ -25,7 +25,7 @@ import {
   maestroScriptSourceBundleFor,
   replayScriptSourceBundleFor,
 } from '../../../__tests__/test-utils/replay-script-source.ts';
-import { REPLAY_SCRIPT_SOURCE_REQUIRED_MESSAGE } from '../../../replay/script-source-bundle.ts';
+import { REPLAY_SCRIPT_SOURCE_REQUIRED_MESSAGE } from '../../replay-script-source.ts';
 import {
   captureSnapshotThroughLegacyDispatchFixture,
   legacyDispatchCapture,
